ajout de clientHandler
